Test target groups being back in port mappings.
The issue with port mappings is that it is a good place to put the
target group as they are linked to the load balancer but the issue
with this is the LB ARN can't be passed to the task.

This should split them in a way that allows them to be specified
together without breaking the task.