Terraform provides no clean way, ugly it is.
There is not a nice way in Terraform to keep the container ports and
target groups together so they are just going to have to be passed as
two objects and let the user do the linking between them.

This is ugly an inelegant but this is how the tool provided works.